// resources/js/api/generated/models/statsResponseData.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { EventTypeCount } from './eventTypeCount';
import type { StatsResponseDataEventsByCategory } from './statsResponseDataEventsByCategory';
import type { StatsResponseDataEventsBySeverity } from './statsResponseDataEventsBySeverity';

export interface StatsResponseData {
  events_by_category: StatsResponseDataEventsByCategory;
  events_by_severity: StatsResponseDataEventsBySeverity;
  events_last_24_hours: number;
  events_last_7_days: number;
  failed_events: number;
  recent_event_types: EventTypeCount[];
  total_events: number;
}